requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
alembic
email-validator
asyncpg
